+++ client/src/sections/index.ts
@@ -3,3 +3,4 @@ export * from "./TodaysImage";
 export * from "./Login";
 export * from "./NotFound";
 export * from "./User";
+export * from "./AppHeader";
